+++ packages/calcite-components/src/components/action-bar/action-bar.e2e.ts
@@ -76,6 +76,37 @@ describe("calcite-action-bar", () => {
     );
   });
 
+  describe("messageOverrides", () => {
+    it("should honor expandLabel and collapseLabel", async () => {
+      const page = await newE2EPage();
+
+      await page.setContent("<calcite-action-bar></calcite-action-bar>");
+      await page.waitForChanges();
+
+      const actionBar = await page.find("calcite-action-bar");
+
+      const expandLabel = "Open me up";
+      const collapseLabel = "Close me down";
+
+      actionBar.setProperty("messageOverrides", {
+        expandLabel,
+        collapseLabel,
+      });
+      await page.waitForChanges();
+
+      const expandAction = await page.find("calcite-action-bar >>> #expand-toggle");
+
+      expect(expandAction).not.toBeNull();
+
+      expect(await expandAction.getProperty("label")).toBe(expandLabel);
+
+      actionBar.setProperty("expanded", true);
+      await page.waitForChanges();
+
+      expect(await expandAction.getProperty("label")).toBe(collapseLabel);
+    });
+  });
+
   describe("expand functionality", () => {
     it("should not modify actions within an action-menu", async () => {
       const page = await newE2EPage({
@@ -110,7 +141,7 @@ describe("calcite-action-bar", () => {
 
       await page.waitForChanges();
 
-      const expandAction = await page.find("calcite-action-bar >>> calcite-action");
+      const expandAction = await page.find("calcite-action-bar >>> #expand-toggle");
 
       expect(expandAction).not.toBeNull();
     });