+++ packages/calcite-components/src/components/date-picker/date-picker.stories.ts
@@ -2,18 +2,14 @@ import { boolean, createBreakpointStories, modesDarkDefault } from "../../../.st
 import { html } from "../../../support/formatting";
 import { locales, defaultLocale } from "../../utils/locale";
 import { ATTRIBUTES } from "../../../.storybook/resources";
+import { DatePicker } from "./date-picker";
 const { scale } = ATTRIBUTES;
 
-interface DatePickerArgs {
+interface DatePickerStoryArgs extends Pick<DatePicker, "min" | "max" | "range" | "scale" | "value"> {
   dir: string;
   lang: string;
-  max: string;
-  min: string;
   nextMonthLabel: string;
   prevMonthLabel: string;
-  range: boolean;
-  scale: string;
-  value: string;
 }
 
 export default {
@@ -48,7 +44,7 @@ export default {
   },
 };
 
-export const simple = (args: DatePickerArgs): string => html`
+export const simple = (args: DatePickerStoryArgs): string => html`
   <div style="width: 400px">
     <calcite-date-picker
       dir="${args.dir}"