+++ packages/calcite-components/src/components/functional/ExpandToggle.tsx
@@ -9,6 +9,8 @@ interface ExpandToggleProps {
   expanded: boolean;
   expandText: string;
   collapseText: string;
+  expandLabel: string;
+  collapseLabel: string;
   el: HTMLElement;
   position: Position;
   tooltip?: HTMLCalciteTooltipElement;
@@ -67,6 +69,8 @@ export const ExpandToggle: FunctionalComponent<ExpandToggleProps & JSXBase.HTMLA
   expanded,
   expandText,
   collapseText,
+  expandLabel,
+  collapseLabel,
   toggle,
   el,
   position,
@@ -77,6 +81,7 @@ export const ExpandToggle: FunctionalComponent<ExpandToggleProps & JSXBase.HTMLA
   const rtl = getElementDir(el) === "rtl";
 
   const text = expanded ? collapseText : expandText;
+  const label = expanded ? collapseLabel : expandLabel;
   const icons = [ICONS.chevronsLeft, ICONS.chevronsRight];
 
   if (rtl) {
@@ -90,6 +95,8 @@ export const ExpandToggle: FunctionalComponent<ExpandToggleProps & JSXBase.HTMLA
   const actionNode = (
     <calcite-action
       icon={expanded ? expandIcon : collapseIcon}
+      id="expand-toggle"
+      label={label}
       onClick={toggle}
       ref={(referenceElement): HTMLCalciteActionElement =>
         setTooltipReference({ tooltip, referenceElement, expanded, ref })